+++ django_fsm/signals.py
@@ -2,5 +2,5 @@ from __future__ import annotations
 
 from django.db.models.signals import ModelSignal
 
-pre_transition = ModelSignal()
-post_transition = ModelSignal()
+pre_transition: ModelSignal = ModelSignal()
+post_transition: ModelSignal = ModelSignal()